refactor: simplify DeploymentState
